Create Source Code
The whole computational process will be analogous to evolution(look at the flowchart). Every new generation of solutions will be created
solely from its ancestors using a more probabilistic approach . Every solution will have its own fitness factor, which is basically a function of number of time a product(or a particular association of products) appears in the transactions database. The chromosomes, whose fitness factor exceeds the computational requirement, will the only ones that are allowed to survive and produce offspring for the next generation. The fitness is calculated by doing certain mathematical operations on the solution string. First, the frequency of each product is taken in a frequency table. These frequencies are divided by the total number of transactions for each case in order to bring frequency
where they can be easily compared in relative manner. The fitness of a given binary string(chromosome) is found by taking the average of relative frequencies of all the live bits in a given chromosome . The dead bits are ignored, because our aim is not only to find the bigger associations , but the smaller ones too. This process will be re-iterated for each generation ,returning a new population every time. The size and mathematical nature of these chromosomes in the most recent generation depends heavily on the cutoff ration specified by the user. This cutoff ratio is number which species the percentage of transaction. that a particular association has to be part in order to qualify as fit solution. Every iteration of fitness evaluation of the population member is followed by a selection procedure. The chromosomes whose fitness exceeds the cutoff ration qualified to participate in the reproduction events of the current iteration(generation). Reproduction, in the context of Genetic Algorithms, refers to concatenation of string in a selective manner, with option shuffling to produce a progeny string which has the same length as that of the parents. The progeny will have a concatenation of partial traits from both its parents. This can be achieved by one point crossover .